agent-paperpal/backend/requirements.txt
# backend/requirements.txt
# Agent Paperpal — Python Dependencies

# ── Web Framework ───────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ── Database ────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# ── Redis / Caching ────────────────────────────────────────────────────────
redis[hiredis]==5.2.1

# ── Celery ──────────────────────────────────────────────────────────────────
celery[redis]==5.4.0

# ── Validation & Settings ──────────────────────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# ── Authentication ──────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── LLM / AI ───────────────────────────────────────────────────────────────
anthropic==0.42.0
langgraph==0.2.60
langchain-core==0.3.29
langchain-anthropic==0.3.4

# ── NLP ─────────────────────────────────────────────────────────────────────
spacy==3.8.3

# ── Document Processing ────────────────────────────────────────────────────
python-docx==1.1.2
pypdf==5.1.0
python-pptx==1.0.2

# ── AWS / S3 ────────────────────────────────────────────────────────────────
boto3==1.35.86
aioboto3==13.3.0

# ── Logging & Monitoring ───────────────────────────────────────────────────
structlog==24.4.0

# ── Testing ─────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1

# ── Linting ─────────────────────────────────────────────────────────────────
ruff==0.8.6
